PWN-4272 - Rename select package to biometrics
