+++ app/src/main/java/org/p2p/wallet/auth/ui/pin/biometrics/BiometricsContract.kt
@@ -1,10 +1,10 @@
-package org.p2p.wallet.auth.ui.pin.select
+package org.p2p.wallet.auth.ui.pin.biometrics
 
 import org.p2p.wallet.common.mvp.MvpPresenter
 import org.p2p.wallet.common.mvp.MvpView
 import javax.crypto.Cipher
 
-interface TouchIdContract {
+interface BiometricsContract {
 
     interface View : MvpView {
         fun showBiometricDialog(cipher: Cipher)
